remove more crap
